자바 : day2 오후수업

+++ java/src/day2/variable/CastEx.java
@@ -14,7 +14,8 @@ public class CastEx {
 		
 		System.out.println(pi);
 		System.out.println(num2);
-		System.out.println(1/2);
+		//1/2는 문법에는 문제가 없지만 원하는 결과를 얻기 위해서 강제형변환을 함.
+		System.out.println((double)1/2);
 
 	}
